requirements.txt
absl-py==0.2.2
astor==0.6.2
backcall==0.1.0
bleach==2.1.3
bottle==0.12.13
cached-property==1.4.3
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
cycler==0.10.0
decorator==4.3.0
entrypoints==0.2.3
Flask==1.0.2
flickrapi==2.4.0
gast==0.2.0
grpcio==1.12.1
gunicorn==19.8.1
h5py==2.8.0
html5lib==1.0.1
idna==2.7
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-client==5.2.3
jupyter-core==4.4.0
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.5.0
numpy==1.14.5
oauthlib==2.1.0
pandocfilters==1.4.2
parso==0.2.1
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
protobuf==3.6.0
Pygments==2.2.0
pymongo==3.6.1
pynder==0.0.13
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
pywinpty==0.5.4
PyYAML==3.12
pyzmq==17.0.0
requests==2.19.1
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
scikit-learn==0.19.1
scipy==1.1.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
tensorboard==1.8.0
tensorflow==1.8.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wincertstore==0.2
